requirements.txt
# Core dependencies
requests==2.31.0
python-dotenv==1.0.1
firecrawl>=0.1.0

# Relevance filtering and sentiment analysis
google.generativeai==0.3.2

# Data storage in Google Sheets
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0

# Optional dependencies
#pandas>=2.0.0  # For data analysis
#matplotlib>=3.5.0  # For visualization

beautifulsoup4==4.10.0
scrapingbee==1.0.0